Added games module, comments, default command
A module can now register default command
for a channel or nick. The default command
will be used if no other command is found.
The games module uses this to get number
guesses from a user without having to prefix
every guess with something like ".guess".
Modules are responsible for clearing the
default command when they're done.