Closes #1 - Added wiretapping for modules to save IRC logs
Meetbot needs to keep a record of messsages sent and received
for a raw meeting log. A custom connection object is used
that intercepts Privmsgs and Actions to handle messages
sent from the bot and custom handlers are registered with the
real irc connection object to handle incoming messages and
actions. Meetbot subscribes a meeting's tap method when a meeting
is started and unregisters it when the meeting is finished.