Added plugin init and shutdown hooks, added Dandelion module
The Dandelion module is untested. It probably doesn't work
yet. I'll test it tomorrow.

The plugins.RegisterInit() and RegisterShutdown() allow modules
to register functions that are ran after the core IRC bot and
msgbus are started and when the application is shutting down
normally. The init functions are given the application config
and IRC bot connection.