Fully implemented message bug and plugin loading
The Dandelion and counter plugins are ported to the new
framework. Meetbot and games will require a bit more work
as each one takes commands and meetbot needs to keep
track of posts written by every user in a channel.

All modules for the message bus have been ported and tested.