Added LibreNMS alert routing based on sysContact
Under [routes.librenms.settings.routes], a table
of sysContact info to channel name can be configured
to route alerts. When an alert comes in and the routes
table exists, the keys will be checked against the sysContact
info from LibreNMS. If the key is contained anywhere in
sysContact, the alert will be sent to that channel.
The check is not an exact match. For example, if the key
is "alerts@example.com" and the sysContact is
"Alerts <alerts@example.com>" it will still match. This means
the key doesn't have to be an email address. It can be any
string in sysContact. There's a special case for the key "*".
The key "*" will always match. This means all alerts will be
posted to that channel regardless of the sysContact.