Added method to register websocket event handlers
Handlers can be registered against a specific channel or "*" to
receive events in all channels. Handlers are mapped to which
events they want to receive. A single handler can be registered
for multiple event types but only one channel. The registeration
function can be called multiple times with the same handler to
listen on multiple channels.